fix cicd

+++ .github/scripts/gen_coverage_report.sh
@@ -11,7 +11,13 @@ git remote -v
 
 git fetch
 git checkout coverage
-git add --force index.html
-git commit -m "[ci skip] Update coverage"
-git remote set-url --push origin https://tomoyane:${GITHUB_TOKEN}@github.com/tomoyane/http-continuous-benchmarking.git
-git push origin HEAD:coverage --force
+
+is_changed=$(git status | grep "nothing to commit" | wc -l)
+if [ "$is_changed" -eq 1 ]; then
+  git add --force index.html
+  git commit -m "[ci skip] Update coverage"
+  git remote set-url --push origin https://tomoyane:${GITHUB_TOKEN}@github.com/tomoyane/http-continuous-benchmarking.git
+  git push origin HEAD:coverage --force
+else
+  echo "Not need to update coverage"
+fi
